\name{Rdroolsjars}
\alias{Rdroolsjars}
\alias{Rdroolsjars}
\docType{package}
\title{
External Jars Required for Package 'Rdrools'
}

\description{
The 'Rdroolsjars' package collects all the external jars required for the 'Rdrools' package. These external jars are quite large in size (10MB) but have a very slow release cycle. By separating the Java and R development, the storage footprint on CRAN is reduced. By using rJava, the Java layer communicates with the R layer and provides the functionality of using Drools in R.

}
\details{
\tabular{ll}{
Package: \tab Rdroolsjars\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2017-03-22\cr
License: \tab Apache License 2.0\cr
}
}
\author{
Ashwin Raaghav, SMS Chauhan

Maintainer: Ashwin Raaghav <ashraaghav@gmail.com>

}


